feat: INFRA-567 Renamed previous template directory to fluxcd to reflect more specialized use-case
