+++ fluxcd/fluxcd.go
@@ -1,4 +1,4 @@
-package template
+package fluxcd
 
 import (
 	"path"
